@@ -5,7 +5,7 @@ author: chinabrant <sjwu1234@gmail.com>
 homepage: https://github.com/chinabrant/popup_menu
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
